python-code/api/requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
openai==1.55.3
pinecone==5.3.1
runpod==1.7.1
urllib3==1.26.6